BREAKING-CHANGE: moved to monorepo architecture
