+++ .gitignore
@@ -4,3 +4,7 @@
 
 # dependencies
 node_modules/
+
+# IDE
+.idea/
+.vscode/
