+++ packages/notion/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@hugo/notion",
+  "private": true,
+  "version": "1.0.0",
+  "main": "index.ts",
+  "author": {
+    "name": "Hugo Ventura",
+    "url": "https://x.com/hugovntr"
+  },
+  "license": "ISC",
+  "dependencies": {
+    "@notionhq/client": "^2.2.11",
+    "dotenv": "^16.3.1",
+    "typescript": "5.1.6"
+  }
+}
